Backend/requirements.txt
psycopg2-binary  # PostgreSQL adapter for Python
python-dotenv    # Loads .env files

pip freeze > requirements.txt # - put this in the terminal to downlad everything at once

# for the API section

pip install supabase
pip install python-dotenv
pip install flask

